split handler into handler, tick
